requirements.txt
# Core dependencies
fastapi
uvicorn
pydantic
pydantic-settings

# LLM clients
openai
groq
google-genai

# Reddit API
asyncpraw

# Vector storage
chromadb

# Web scraping
requests
beautifulsoup4
firecrawl-py

# Utilities
python-dotenv